cv_survtmle_internal

This function estimates the marginal cumulative incidence for failures of specified types using cross-validated targeted minimum loss-based estimation. More description to come...

cv_survtmle_internal(ftime, ftype, trt, adjustVars, t0 = max(ftime),
  incidence = TRUE, survtmleFits, validFolds, returnIC = TRUE,
  returnModels = TRUE, ftypeOfInterest = "all", trtOfInterest = "all",
  method = "hazard", bounds = NULL, verbose = TRUE,
  tol = 1/(length(ftime)), maxIter = 100)

Arguments

ftime

A numeric vector of failure times. Missing values are not supported. Right censored observations should have a non-missing value for ftime while ftype should be set to 0.

ftype

A numeric vector indicating the type of failure with type 0 reserved for right censored observations. Each unique value will be treated as an (unordered) separate type of failure.

trt

A numeric vector indicating observed treatment assignment. Each unique value will be treated as an (unordered) separate type of treatment. Currently, only two unique values of trt are supported.

adjustVars

A data.frame of adjustment variables that will be used in estimating the conditional treatment, censoring, and failure (hazard or conditional mean) probabilities.

t0

The time at which to return cumulative incidence estimates. By default this is set to max(ftime).

incidence

If TRUE the function return estimates of cumulative incidence. If FALSE the function returns esimtates of survival.

survtmleFits

A list of objects of class survtmle fit on train samples of data

validFolds

A list of vectors indicating which observations were in the validation fold for each of the survtmleFits.

returnIC

A boolean indicating whether to return vectors of influence curve estimates. These are needed for some post-hoc comarisons, so it is recommended to leave as TRUE (the default) unless the user is sure these estimates will not be needed later.

returnModels

A boolean indicating whether to return the SuperLearner or glm objects used to estimate the nuisance parameters. Must be set to TRUE if the user plans to use calls to timepoints to obtain estimates at times other than t0. See ?timepoints for more information.

ftypeOfInterest

An input specifying what failure types to compute estimates of incidence for. The default value is "all", which computes estimates for values unique(ftype). Can alternatively be set to a vector of values found in ftype.

trtOfInterest

An input specifying which levels of trt are of interest. The default value is "all", which computes estimates for values unique(trt). Can alternatively be set to a vector of values found in trt.

method

A character specification of how the targeted minimum loss-based estimators should be computer, either "mean" or "hazard". The "mean" specification uses a closed-form targeted minimum loss-based estimation based on the G-computation formula of Bang and Robins (2005). The "hazard" specification uses an iteratively algorithm based on cause-specific hazard functions. The latter specification has no guarantee of convergence in finite samples. The convergence can be influenced by the stopping criteria specified in the tol. Future versions may implement a closed form version of this hazard-based estimator.

bounds

A list of bounds... XXX NEED MORE DESCRIPTION HERE XXX

verbose

A boolean indicating whether the function should print messages to indicate progress.

tol

The stopping criteria when method="hazard". The algorithm will continue performing targeting updates to the initial estimators until the empirical mean of the efficient influence function is smaller than tol. The default (1/length(ftime)) is a sensible value. Larger values can be used in situations where convergence of the algorithm is an issue; however, this may result in large finite-sample bias.

maxIter

A maximum number of iterations for the algorithm when method="hazard". The algorithm will iterate until either the empirical mean of the efficient influence function is smaller than tol or until maxIter iterations have been completed.

Value

An object of class cv_survtmle.

call The call to cv_survtmle.

est A numeric vector of point estimates -- one for each combination of ftypeOfInterest and trtOfInterest.

var A covariance matrix of all the point estimates

meanIC The empirical mean of the efficient influence function at the estimated, fluctuated nuisance parameters. If all goes well, each value should be small. If method="hazard" then each value should be less than tol. If codemethod="mean" then should be on the order of 1e-7.

ic The efficient influence function at the estimated, fluctuated nuisance parameters, evaluated on each of the observations (summed over all times). These may be used to perform post-hoc comparisons. More details coming.

ftimeMod If returnModels=TRUE the fit object(s) for the call to glm or SuperLearner for the outcome regression models. If method="mean" this will be a list of length length(ftypeOfInterest) each of length t0 (one regression for each failure type and for each time point). If method="hazard" this will be a list of length length(ftypeOfInterest) with one model corresponding with the (pseudo-)hazard for each cause of failure. If returnModels=FALSE, this will equal NULL.

ctimeMod If returnModels=TRUE the fit object for the call to glm or SuperLearner for the censoring hazard regression model. If returnModels=FALSE, this will equal NULL.

trtMod If returnModels=TRUE the fit object for the call to glm or SuperLearner for the conditioanl probability of trt regression model. If returnModels=FALSE, this will equal NULL.

t0 The timepoint the function was evaluated at.