hazard.tmle

This function estimates the marginal cumulative incidence for failures of specified types using targeted minimum loss-based estimation based on the initial estimates of the cause-specific hazard functions for failures of each type. The function is called by survtmle whenever method="hazard" is specified. However, power users could, in theory, make calls directly to this function.

hazard.tmle(ftime, ftype, trt, t0, incidence = TRUE, adjustVars = NULL,
  SL.ftime = NULL, SL.ctime = NULL, SL.trt = NULL, glm.ftime = NULL,
  glm.ctime = NULL, glm.trt = "1", returnIC = TRUE,
  returnModels = FALSE, ftypeOfInterest = "all", trtOfInterest = "all",
  method = "hazard", bounds = NULL, verbose, tol = 1/(length(ftime)),
  maxIter = 100, ...)

Arguments

ftime

A numeric vector of failure times. Missing values or zero values are not supported. Right censored observations should have a non-missing value for ftime while ftype should be set to 0.

ftype

A numeric vector indicating the type of failure with type 0 reserved for right censored observations. Each unique value will be treated as an (unordered) separate type of failure.

trt

A numeric vector indicating observed treatment assignment. Each unique value will be treated as an (unordered) separate type of treatment. Currently, only two unique values of trt are supported.

t0

The time at which to return cumulative incidence estimates. By default this is set to max(ftime).

incidence

If TRUE the function return estimates of cumulative incidence. If FALSE the function returns esimtates of survival.

adjustVars

A data.frame of adjustment variables that will be used in estimating the conditional treatment, censoring, and failure (hazard or conditional mean) probabilities.

SL.ftime

A character vector or list specification to be passed to the SL.library argument in the call to SuperLearner for the outcome regression (either cause-specific hazards or condtional mean). See ?SuperLearner for more information on how to specify valid SuperLearner libraries. It is expected that the wrappers used in the library will play nicely with the input variables, which will be called "trt" and names(adjustVars).

SL.ctime

A character vector or list specification to be passed to the SL.library argument in the call to SuperLearner for the estimate of the conditional hazard for censoring. It is expected that the wrappers used in the library will play nicely with the input variables, which will be called "trt" and names(adjustVars).

SL.trt

A character vector or list specification to be passed to the SL.library argument in the call to SuperLearner for the estimate of the conditional probabiltiy of treatment. It is expected that the wrappers used in the library will play nicely with the input variables, which will be names(adjustVars).

glm.ftime

A character specification of the right-hand side of the equation passed to the formula option of a call to glm for the outcome regression (either cause-specific hazards or condtional mean). Ignored if SL.ftime != NULL. Use "trt" to specify the treatment in this formula (see examples). The formula can additionally include any variables found in names(adjustVars).

glm.ctime

A character specification of the right-hand side of the equation passed to the formula option of a call to glm for the estimate of the conditional hazard for censoring. Ignored if SL.ctime != NULL. Use "trt" to specify the treatment in this formula (see examples). The formula can additionally include any variables found in names(adjustVars).

glm.trt

A character specification of the right-hand side of the equation passed to the formula option of a call to glm for the estimate of the conditional probabiltiy of treatment. Ignored if SL.trt != NULL. By default set to "1", corresponding to using empirical estimates of each value of trt. The formula can include any variables found in names(adjustVars).

returnIC

A boolean indicating whether to return vectors of influence curve estimates. These are needed for some post-hoc comarisons, so it is recommended to leave as TRUE (the default) unless the user is sure these estimates will not be needed later.

returnModels

A boolean indicating whether to return the SuperLearner or glm objects used to estimate the nuisance parameters. Must be set to TRUE if the user plans to use calls to timepoints to obtain estimates at times other than t0. See ?timepoints for more information.

ftypeOfInterest

An input specifying what failure types to compute estimates of incidence for. The default value is "all", which computes estimates for values unique(ftype). Can alternatively be set to a vector of values found in ftype.

trtOfInterest

An input specifying which levels of trt are of interest. The default value is "all", which computes estimates for values unique(trt). Can alternatively be set to a vector of values found in trt.

bounds

A list of bounds... XXX NEED MORE DESCRIPTION HERE XXX

verbose

A boolean indicating whether the function should print messages to indicate progress.

tol

The stopping criteria for the fluctuation step. The algorithm will continue performing targeting updates to the initial estimators until the empirical mean of the efficient influence function is smaller than tol. The default (1/length(ftime)) is a sensible value. Larger values can be used in situations where convergence of the algorithm is an issue; however, this may result in finite-sample bias.

maxIter

A maximum number of iterations for the algorithm when method="hazard". The algorithm will iterate until either the empirical mean of the efficient influence function is smaller than tol or until maxIter iterations have been completed.

...

Other arguments. Not currently used.

Value

est A numeric vector of point estimates -- one for each combination of ftypeOfInterest and trtOfInterest.

var A covariance matrix of all the point estimates

meanIC The empirical mean of the efficient influence function at the estimated, fluctuated nuisance parameters. If all goes well, each value should be less than tol.

ic The efficient influence function at the estimated, fluctuated nuisance parameters, evaluated on each of the observations (summed over all times). These may be used to perform post-hoc comparisons. More details coming.

ftimeMod If returnModels=TRUE the fit object(s) for the call to glm or SuperLearner for the outcome regression models. If method="mean" this will be a list of length length(ftypeOfInterest) each of length t0 (one regression for each failure type and for each time point). If method="hazard" this will be a list of length length(ftypeOfInterest) with one model corresponding with the (pseudo-)hazard for each cause of failure. If returnModels=FALSE, this will equal NULL.

ctimeMod If returnModels=TRUE the fit object for the call to glm or SuperLearner for the censoring hazard regression model. If returnModels=FALSE, this will equal NULL.

trtMod If returnModels=TRUE the fit object for the call to glm or SuperLearner for the conditioanl probability of trt regression model. If returnModels=FALSE, this will equal NULL.

Examples

## Single failure type examples
# simulate data
set.seed(1234)
n <- 100
trt <- rbinom(n,1,0.5)
adjustVars <- data.frame(W1 = round(runif(n)), W2 = round(runif(n,0,2)))

ftime <- round(1 + runif(n,1,4) - trt + adjustVars$W1 + adjustVars$W2)
ftype <- round(runif(n,0,1))

#' # Fit 1
# fit mean.tmle object with glm estimators for treatment, censoring, and failure
fit1 <- mean.tmle(ftime = ftime, ftype = ftype, trt = trt, adjustVars = adjustVars,
glm.trt = "W1 + W2",
glm.ftime = "trt + W1 + W2", glm.ctime = "trt + W1 + W2",
method="hazard", returnModels = TRUE)

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading

#> Warning: prediction from a rank-deficient fit may be misleading