timepoints

This is a wrapper function for survtmle that takes a fitted survtmle object and computes the TMLE estimated incidence for all times specified in the times argument. For this function to work, the original call to survtmle should have been executed with returnModels = TRUE. This allows the function to be more efficient than repeated calls to survtmle in that timepoints will use fitted censoring (and hazard if method="hazard" was used in the original call) estimates. It is therefore advisable that the vector times used in the call to timepoints not include times beyond the time specified in t0 in the original call to survtmle. This can be ensured be making the original call to survtmle with t0 = max(ftime).

timepoints(object, times, returnModels = FALSE)

Arguments

object

A fitted survtmle object with returnModels=TRUE

times

The times to evaluate incidence.

returnModels

Should the function return fitted glm/super learner models at each timepoint. If set to TRUE, memory issues could arise depending on the number of timepoints specified and the size of the super learner library.

Value

An object of class tp.survtmle with number of entries equal to length(times). Each entry is named "tX", where X denotes a single value of times.

Examples

# simulate data
set.seed(1234)
n <- 100
ftime <- round(runif(n,1,4))
ftype <- round(runif(n,0,2))
trt <- rbinom(n,1,0.5)
adjustVars <- data.frame(W1 = rnorm(n), W2 = rnorm(n))

# fit an initial survtmle object with t0=max(ftime)
fm <- survtmle(ftime = ftime, ftype = ftype, trt = trt, adjustVars = adjustVars,
glm.ftime = "trt + W1 + W2", glm.ctime = "trt + W1 + W2", method="mean", returnModels = TRUE)

# call timepoints to get cumulative incidence estimates at each timepoint
allTimes <- timepoints(object = fm, times = 1:4, returnModels = FALSE)

#> Error in survtmle(ftime = ftime, ftype = ftype, trt = trt, adjustVars = adjustVars,     glm.ftime = "trt + W1 + W2", method = "mean", returnModels = FALSE,     glm.ctime = structure(list(coefficients = structure(c(-1.84079779173557,     -0.615165775080355, -0.422093217869613, 0.202584992054308    ), .Names = c("(Intercept)", "trt", "W1", "W2")), R = structure(c(-3.92768819894934,     0, 0, 0, -1.61372093289878, -1.93238128102824, 0, 0, 0.837691387913871,     -0.183427267790544, -4.22558842491134, 0, -0.808669473789295,     0.264401650628091, -0.299796912574902, 3.51100076632515), .Dim = c(4L,     4L), .Dimnames = list(c("(Intercept)", "trt", "W1", "W2"),         c("(Intercept)", "trt", "W1", "W2"))), rank = 4L, qr = structure(list(        rank = 4L, qraux = c(1.07604541118264, 1.08212650063318,         1.01714791599405, 1.00141344926131), pivot = 1:4, tol = 1e-11), .Names = c("rank",     "qraux", "pivot", "tol"), class = "qr"), family = structure(list(        family = "binomial", link = "logit", linkfun = function (mu)         .Call(C_logit_link, mu), linkinv = function (eta)         .Call(C_logit_linkinv, eta), mu.eta = function (eta)         .Call(C_logit_mu_eta, eta), initialize = expression({            if (NCOL(y) == 1) {                if (is.factor(y))                   y <- y != levels(y)[1L]                n <- rep.int(1, nobs)                y[weights == 0] <- 0                if (any(y < 0 | y > 1))                   stop("y values must be 0 <= y <= 1")                mustart <- (weights * y + 0.5)/(weights + 1)                m <- weights * y                if (any(abs(m - round(m)) > 0.001))                   warning("non-integer #successes in a binomial glm!")            }            else if (NCOL(y) == 2) {                if (any(abs(y - round(y)) > 0.001))                   warning("non-integer counts in a binomial glm!")                n <- y[, 1] + y[, 2]                y <- ifelse(n == 0, 0, y[, 1]/n)                weights <- weights * n                mustart <- (n * y + 0.5)/(n + 1)            }            else stop("for the 'binomial' family, y must be a vector of 0 and 1's\nor a 2 column matrix where col 1 is no. successes and col 2 is no. failures")        }), valideta = function (eta)         TRUE), .Names = c("family", "link", "linkfun", "linkinv",     "mu.eta", "initialize", "valideta"), class = "family"), deviance = 107.805345858316,         aic = 115.805345858316, null.deviance = 113.256396876776,         iter = 5L, df.residual = 159L, df.null = 162L, converged = TRUE,         boundary = FALSE, call = glm(formula = as.formula(ctimeForm),             family = "binomial", data = dataList[[1]][include,                 ]), formula = C ~ trt + W1 + W2, terms = C ~             trt + W1 + W2, offset = NULL, control = structure(list(            epsilon = 1e-08, maxit = 25, trace = FALSE), .Names = c("epsilon",         "maxit", "trace")), method = "glm.fit", contrasts = NULL,         xlevels = structure(list(), .Names = character(0))), .Names = c("coefficients",     "R", "rank", "qr", "family", "deviance", "aic", "null.deviance",     "iter", "df.residual", "df.null", "converged", "boundary",     "call", "formula", "terms", "offset", "control", "method",     "contrasts", "xlevels"), class = c("glm", "lm")), glm.trt = structure(list(        coefficients = structure(0.0800427076726136, .Names = "(Intercept)"),         residuals = structure(c(1.92307692307778, 1.92307692307778,         -2.08333333333233, 1.92307692307778, 1.92307692307778,         1.92307692307778, -2.08333333333233, -2.08333333333233,         -2.08333333333233, 1.92307692307778, 1.92307692307778,         -2.08333333333233, -2.08333333333233, 1.92307692307778,         1.92307692307778, 1.92307692307778, 1.92307692307778,         1.92307692307778, -2.08333333333233, 1.92307692307778,         -2.08333333333233, -2.08333333333233, -2.08333333333233,         -2.08333333333233, -2.08333333333233, -2.08333333333233,         1.92307692307778, -2.08333333333233, 1.92307692307778,         -2.08333333333233, 1.92307692307778, 1.92307692307778,         -2.08333333333233, 1.92307692307778, 1.92307692307778,         -2.08333333333233, 1.92307692307778, -2.08333333333233,         -2.08333333333233, 1.92307692307778, -2.08333333333233,         -2.08333333333233, -2.08333333333233, 1.92307692307778,         1.92307692307778, -2.08333333333233, 1.92307692307778,         1.92307692307778, 1.92307692307778, 1.92307692307778,         1.92307692307778, 1.92307692307778, -2.08333333333233,         -2.08333333333233, -2.08333333333233, 1.92307692307778,         -2.08333333333233, -2.08333333333233, 1.92307692307778,         1.92307692307778, 1.92307692307778, -2.08333333333233,         1.92307692307778, -2.08333333333233, -2.08333333333233,         -2.08333333333233, 1.92307692307778, -2.08333333333233,         -2.08333333333233, 1.92307692307778, -2.08333333333233,         1.92307692307778, -2.08333333333233, 1.92307692307778,         -2.08333333333233, 1.92307692307778, 1.92307692307778,         1.92307692307778, -2.08333333333233, -2.08333333333233,         -2.08333333333233, -2.08333333333233, 1.92307692307778,         1.92307692307778, 1.92307692307778, -2.08333333333233,         -2.08333333333233, -2.08333333333233, 1.92307692307778,         1.92307692307778, 1.92307692307778, -2.08333333333233,         1.92307692307778, -2.08333333333233, 1.92307692307778,         1.92307692307778, 1.92307692307778, -2.08333333333233,         1.92307692307778, -2.08333333333233), .Names = c("1",         "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12",         "13", "14", "15", "16", "17", "18", "19", "20", "21",         "22", "23", "24", "25", "26", "27", "28", "29", "30",         "31", "32", "33", "34", "35", "36", "37", "38", "39",         "40", "41", "42", "43", "44", "45", "46", "47", "48",         "49", "50", "51", "52", "53", "54", "55", "56", "57",         "58", "59", "60", "61", "62", "63", "64", "65", "66",         "67", "68", "69", "70", "71", "72", "73", "74", "75",         "76", "77", "78", "79", "80", "81", "82", "83", "84",         "85", "86", "87", "88", "89", "90", "91", "92", "93",         "94", "95", "96", "97", "98", "99", "100")), fitted.values = structure(c(0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977,         0.51999999999977, 0.51999999999977, 0.51999999999977), .Names = c("1",         "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12",         "13", "14", "15", "16", "17", "18", "19", "20", "21",         "22", "23", "24", "25", "26", "27", "28", "29", "30",         "31", "32", "33", "34", "35", "36", "37", "38", "39",         "40", "41", "42", "43", "44", "45", "46", "47", "48",         "49", "50", "51", "52", "53", "54", "55", "56", "57",         "58", "59", "60", "61", "62", "63", "64", "65", "66",         "67", "68", "69", "70", "71", "72", "73", "74", "75",         "76", "77", "78", "79", "80", "81", "82", "83", "84",         "85", "86", "87", "88", "89", "90", "91", "92", "93",         "94", "95", "96", "97", "98", "99", "100")), effects = structure(c(-0.399893293700079,         0.873426174798747, -1.12817547578172, 0.873426174798747,         0.873426174798747, 0.873426174798747, -1.12817547578172,         -1.12817547578172, -1.12817547578172, 0.873426174798747,         0.873426174798747, -1.12817547578172, -1.12817547578172,         0.873426174798747, 0.873426174798747, 0.873426174798747,         0.873426174798747, 0.873426174798747, -1.12817547578172,         0.873426174798747, -1.12817547578172, -1.12817547578172,         -1.12817547578172, -1.12817547578172, -1.12817547578172,         -1.12817547578172, 0.873426174798747, -1.12817547578172,         0.873426174798747, -1.12817547578172, 0.873426174798747,         0.873426174798747, -1.12817547578172, 0.873426174798747,         0.873426174798747, -1.12817547578172, 0.873426174798747,         -1.12817547578172, -1.12817547578172, 0.873426174798747,         -1.12817547578172, -1.12817547578172, -1.12817547578172,         0.873426174798747, 0.873426174798747, -1.12817547578172,         0.873426174798747, 0.873426174798747, 0.873426174798747,         0.873426174798747, 0.873426174798747, 0.873426174798747,         -1.12817547578172, -1.12817547578172, -1.12817547578172,         0.873426174798747, -1.12817547578172, -1.12817547578172,         0.873426174798747, 0.873426174798747, 0.873426174798747,         -1.12817547578172, 0.873426174798747, -1.12817547578172,         -1.12817547578172, -1.12817547578172, 0.873426174798747,         -1.12817547578172, -1.12817547578172, 0.873426174798747,         -1.12817547578172, 0.873426174798747, -1.12817547578172,         0.873426174798747, -1.12817547578172, 0.873426174798747,         0.873426174798747, 0.873426174798747, -1.12817547578172,         -1.12817547578172, -1.12817547578172, -1.12817547578172,         0.873426174798747, 0.873426174798747, 0.873426174798747,         -1.12817547578172, -1.12817547578172, -1.12817547578172,         0.873426174798747, 0.873426174798747, 0.873426174798747,         -1.12817547578172, 0.873426174798747, -1.12817547578172,         0.873426174798747, 0.873426174798747, 0.873426174798747,         -1.12817547578172, 0.873426174798747, -1.12817547578172        ), .Names = c("(Intercept)", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "", "", "", "", "", "",         "", "", "", "", "", "", "", "", "")), R = structure(-4.99599907758869, .Dim = c(1L,         1L), .Dimnames = list("(Intercept)", "(Intercept)")),         rank = 1L, qr = structure(list(qr = structure(c(-4.99599907758869,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1,         0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1, 0.1        ), .Dim = c(100L, 1L), .Dimnames = list(c("1", "2", "3",         "4", "5", "6", "7", "8", "9", "10", "11", "12", "13",         "14", "15", "16", "17", "18", "19", "20", "21", "22",         "23", "24", "25", "26", "27", "28", "29", "30", "31",         "32", "33", "34", "35", "36", "37", "38", "39", "40",         "41", "42", "43", "44", "45", "46", "47", "48", "49",         "50", "51", "52", "53", "54", "55", "56", "57", "58",         "59", "60", "61", "62", "63", "64", "65", "66", "67",         "68", "69", "70", "71", "72", "73", "74", "75", "76",         "77", "78", "79", "80", "81", "82", "83", "84", "85",         "86", "87", "88", "89", "90", "91", "92", "93", "94",         "95", "96", "97", "98", "99", "100"), "(Intercept)")),             rank = 1L, qraux = 1.1, pivot = 1L, tol = 1e-11), .Names = c("qr",         "rank", "qraux", "pivot", "tol"), class = "qr"), family = structure(list(            family = "binomial", link = "logit", linkfun = function (mu)             .Call(C_logit_link, mu), linkinv = function (eta)             .Call(C_logit_linkinv, eta), variance = function (mu)             mu * (1 - mu), dev.resids = function (y, mu, wt)             .Call(C_binomial_dev_resids, y, mu, wt), aic = function (y,                 n, mu, wt, dev)             {                m <- if (any(n > 1))                   n                else wt                -2 * sum(ifelse(m > 0, (wt/m), 0) * dbinom(round(m *                   y), round(m), mu, log = TRUE))            }, mu.eta = function (eta)             .Call(C_logit_mu_eta, eta), initialize = expression(                {                  if (NCOL(y) == 1) {                    if (is.factor(y))                       y <- y != levels(y)[1L]                    n <- rep.int(1, nobs)                    y[weights == 0] <- 0                    if (any(y < 0 | y > 1))                       stop("y values must be 0 <= y <= 1")                    mustart <- (weights * y + 0.5)/(weights +                       1)                    m <- weights * y                    if (any(abs(m - round(m)) > 0.001))                       warning("non-integer #successes in a binomial glm!")                  }                  else if (NCOL(y) == 2) {                    if (any(abs(y - round(y)) > 0.001))                       warning("non-integer counts in a binomial glm!")                    n <- y[, 1] + y[, 2]                    y <- ifelse(n == 0, 0, y[, 1]/n)                    weights <- weights * n                    mustart <- (n * y + 0.5)/(n + 1)                  }                  else stop("for the 'binomial' family, y must be a vector of 0 and 1's\nor a 2 column matrix where col 1 is no. successes and col 2 is no. failures")                }), validmu = function (mu)             all(is.finite(mu)) && all(mu > 0 & mu < 1), valideta = function (eta)             TRUE, simulate = function (object, nsim)             {                ftd <- fitted(object)                n <- length(ftd)                ntot <- n * nsim                wts <- object$prior.weights                if (any(wts%%1 != 0))                   stop("cannot simulate from non-integer prior.weights")                if (!is.null(m <- object$model)) {                  y <- model.response(m)                  if (is.factor(y)) {                    yy <- factor(1 + rbinom(ntot, size = 1, prob = ftd),                       labels = levels(y))                    split(yy, rep(seq_len(nsim), each = n))                  }                  else if (is.matrix(y) && ncol(y) == 2) {                    yy <- vector("list", nsim)                    for (i in seq_len(nsim)) {                      Y <- rbinom(n, size = wts, prob = ftd)                      YY <- cbind(Y, wts - Y)                      colnames(YY) <- colnames(y)                      yy[[i]] <- YY                    }                    yy                  }                  else rbinom(ntot, size = wts, prob = ftd)/wts                }                else rbinom(ntot, size = wts, prob = ftd)/wts            }), .Names = c("family", "link", "linkfun", "linkinv",         "variance", "dev.resids", "aic", "mu.eta", "initialize",         "validmu", "valideta", "simulate"), class = "family"),         linear.predictors = structure(c(0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136, 0.0800427076726136,         0.0800427076726136, 0.0800427076726136), .Names = c("1",         "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12",         "13", "14", "15", "16", "17", "18", "19", "20", "21",         "22", "23", "24", "25", "26", "27", "28", "29", "30",         "31", "32", "33", "34", "35", "36", "37", "38", "39",         "40", "41", "42", "43", "44", "45", "46", "47", "48",         "49", "50", "51", "52", "53", "54", "55", "56", "57",         "58", "59", "60", "61", "62", "63", "64", "65", "66",         "67", "68", "69", "70", "71", "72", "73", "74", "75",         "76", "77", "78", "79", "80", "81", "82", "83", "84",         "85", "86", "87", "88", "89", "90", "91", "92", "93",         "94", "95", "96", "97", "98", "99", "100")), deviance = 138.469393417992,         aic = 140.469393417992, null.deviance = 138.469393417992,         iter = 3L, weights = structure(c(0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671, 0.249600067832671,         0.249600067832671, 0.249600067832671), .Names = c("1",         "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12",         "13", "14", "15", "16", "17", "18", "19", "20", "21",         "22", "23", "24", "25", "26", "27", "28", "29", "30",         "31", "32", "33", "34", "35", "36", "37", "38", "39",         "40", "41", "42", "43", "44", "45", "46", "47", "48",         "49", "50", "51", "52", "53", "54", "55", "56", "57",         "58", "59", "60", "61", "62", "63", "64", "65", "66",         "67", "68", "69", "70", "71", "72", "73", "74", "75",         "76", "77", "78", "79", "80", "81", "82", "83", "84",         "85", "86", "87", "88", "89", "90", "91", "92", "93",         "94", "95", "96", "97", "98", "99", "100")), prior.weights = structure(c(1,         1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,         1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,         1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,         1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,         1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,         1, 1, 1, 1, 1, 1, 1, 1, 1), .Names = c("1", "2", "3",         "4", "5", "6", "7", "8", "9", "10", "11", "12", "13",         "14", "15", "16", "17", "18", "19", "20", "21", "22",         "23", "24", "25", "26", "27", "28", "29", "30", "31",         "32", "33", "34", "35", "36", "37", "38", "39", "40",         "41", "42", "43", "44", "45", "46", "47", "48", "49",         "50", "51", "52", "53", "54", "55", "56", "57", "58",         "59", "60", "61", "62", "63", "64", "65", "66", "67",         "68", "69", "70", "71", "72", "73", "74", "75", "76",         "77", "78", "79", "80", "81", "82", "83", "84", "85",         "86", "87", "88", "89", "90", "91", "92", "93", "94",         "95", "96", "97", "98", "99", "100")), df.residual = 99L,         df.null = 99L, y = structure(c(1, 1, 0, 1, 1, 1, 0, 0,         0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0,         1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1,         1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0,         1, 0, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0,         0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0,         1, 0), .Names = c("1", "2", "3", "4", "5", "6", "7",         "8", "9", "10", "11", "12", "13", "14", "15", "16", "17",         "18", "19", "20", "21", "22", "23", "24", "25", "26",         "27", "28", "29", "30", "31", "32", "33", "34", "35",         "36", "37", "38", "39", "40", "41", "42", "43", "44",         "45", "46", "47", "48", "49", "50", "51", "52", "53",         "54", "55", "56", "57", "58", "59", "60", "61", "62",         "63", "64", "65", "66", "67", "68", "69", "70", "71",         "72", "73", "74", "75", "76", "77", "78", "79", "80",         "81", "82", "83", "84", "85", "86", "87", "88", "89",         "90", "91", "92", "93", "94", "95", "96", "97", "98",         "99", "100")), converged = TRUE, boundary = FALSE, model = structure(list(            thisY = c(1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 0, 0,             1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0,             1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1,             1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1, 0,             0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 0,             0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 0, 1, 1, 1, 0,             1, 0)), .Names = "thisY", terms = thisY ~ 1, row.names = c(NA,         100L), class = "data.frame"), call = stats::glm(formula = stats::as.formula(paste0("thisY ~ ",             glm.trt)), family = "binomial", data = adjustVars),         formula = thisY ~ 1, terms = thisY ~ 1, data = structure(list(            W1 = c(-0.377237647496313, 0.0976194631092374, 1.63874464512173,             -0.875592474189569, 0.121759998641046, 1.36213066084498,             -0.23462108692271, -1.05338280794309, -0.869783605635912,             -0.390127029647055, -0.847350073129572, -0.260639392099399,             -0.414419706341893, -0.183050797719434, 0.407056097880578,             0.624633128202147, 1.67820574370471, -0.0686936535649907,             -0.320839912658177, 1.47100571704891, 1.7043293978271,             0.0432440375575291, -0.332657319133519, -1.8222354175631,             1.41126239874678, -0.837582434076861, -1.12376279439532,             3.04376588609529, 0.235021308250671, -0.0332586101133057,             -2.7322195229558, -0.0997905884418961, 0.976031734922396,             0.413868915451097, 0.912322161610113, 1.98373220068438,             1.16910851401363, -0.50873701541522, 0.704180178465512,             -0.198416273822079, -0.538070788884863, -2.85575865501923,             -0.789646852263761, 0.487814635163685, 2.16803253951933,             0.500694614280786, 0.62021020366732, -0.965903210133575,             0.162654708118265, -2.07823754188738, 0.485226820569252,             0.696768778564913, 0.185513915583057, 0.700733515544461,             0.311681028661359, 0.760462361967838, 1.84246362620766,             1.1123628412626, 0.0326639575014441, -1.11444896479736,             0.418057822385083, -0.400235237343163, 1.49349310261748,             -1.60708093984972, -0.415751788401515, 0.42200837321742,             -0.151736536534977, -0.60615111526422, -0.304721068966714,             0.629536099884472, 0.895171980275539, 0.660212631820405,             2.27348352044748, 1.17349757263239, 0.287709728313787,             -0.659770093821306, 2.91914013071762, 0.677415500438328,             -0.684320344136007, 0.186492083080971, -0.324393300483657,             -0.274704218225806, -0.933503340589868, 0.116845344986082,             0.319160238648117, -1.07754212275943, -3.23315213292314,             -0.254874652654534, 0.0295178303214797, 0.594273774110513,             0.0591351681787969, 0.413398894737046, -1.09777217457042,             0.711175257270441, 0.718888729854143, 0.251651069028968,             1.35727443615177, 0.404468471278607, 0.264364269837939,             0.268043904143697), W2 = c(0.436930577039483, 1.06012390548736,             0.452190396528715, 0.663198615687065, -1.13637355399084,             -0.370497517092117, 1.47696958979247, -1.22390375051376,             0.258068387396106, 0.405002805433516, 0.97580332180945,             -0.348876736539909, 0.158625439491262, -1.76325506654115,             0.338596047099905, -0.666565029559273, -0.238646624327644,             -1.18776528245719, 0.384935321768578, 0.66657951565448,             -0.304613889457735, 1.8250110640863, 0.670559370567588,             0.948632573255153, 2.04940300167091, -0.651113609237535,             0.808619272846293, 0.986580613954563, -0.00617079609070003,             0.319052357931912, -1.01182190274184, 0.470167547726389,             -0.700970331629097, 0.813682862526491, -0.811430783992772,             0.31939748709946, -0.846522653163624, -0.245763179128733,             -1.55285901110546, 0.128434032925753, 0.985443389404868,             0.183247523067191, -1.76622921343046, -0.620533696615944,             1.65604303699874, 1.80980538558388, -1.17503676801214,             -0.366703258827641, 0.353625448927287, 0.319156220867898,             -0.579956989019402, -0.953278701946907, -0.179428586945621,             1.0098082149443, 0.0236266149214088, -0.649028219583901,             -0.504374222048191, 1.61439149545836, -0.446959811255663,             0.763176761113315, 1.4717186861858, 0.443664903687416,             -0.421721870070189, -0.0400016262060736, -0.492279968227887,             1.22771711731275, -0.149553565244392, 1.54998338398246,             -0.561612535381314, -0.647117250632361, 0.143132159050427,             0.0241886479910512, -0.504451524580869, -1.58139680993034,             0.030066421255418, -0.716576699070701, 1.08261095547612,             -0.952685454939293, 1.12648272900916, -0.649043017379116,             0.292470078406726, 0.898702724402737, -0.518742356791874,             0.554438552345125, -0.0879736711840128, -1.13521292940752,             -0.270079610999902, 1.61978987753494, -0.21413117369664,             -0.817782462924957, -0.0540229240816071, 0.330141609857869,             0.955324614657399, 1.14395987779244, 0.100522398375063,             1.16457524565738, -0.764259947799148, -2.3445134046903,             -0.471683425529536, -0.51585548950739)), .Names = c("W1",         "W2"), row.names = c(NA, -100L), class = "data.frame"),         offset = NULL, control = structure(list(epsilon = 1e-08,             maxit = 25, trace = FALSE), .Names = c("epsilon",         "maxit", "trace")), method = "glm.fit", contrasts = NULL,         xlevels = NULL), .Names = c("coefficients", "residuals",     "fitted.values", "effects", "R", "rank", "qr", "family",     "linear.predictors", "deviance", "aic", "null.deviance",     "iter", "weights", "prior.weights", "df.residual", "df.null",     "y", "converged", "boundary", "model", "call", "formula",     "terms", "data", "offset", "control", "method", "contrasts",     "xlevels"), class = c("glm", "lm")), t0 = 1L): object 'ftime' not found

# look at results for time 1
class(allTimes$t1)

#> Error in eval(expr, envir, enclos): object 'allTimes' not found

allTimes$t1

#> Error in eval(expr, envir, enclos): object 'allTimes' not found

# look at results for time 2 
allTimes$t2

#> Error in eval(expr, envir, enclos): object 'allTimes' not found