Convert Short Form Data to List of Wide Form Data

The function takes a data.frame of short format right-censored failure times and reshapes the long format into the wide format needed for calls to both mean_tmle and hazard_tmle. The list returned will have a number of entries equal to length(trtOfInterest) + 1. The first will have number of rows for each observation equal to the ftime corresponding to that observation. The subsequent entries will have t0 rows for each observation and will set trt column equal to each value of trtOfInterest in turn.

makeDataList(dat, J, ntrt, uniqtrt, t0, bounds = NULL, ...)

Arguments

dat
The short form data.frame

J
The unique values of ftype passed to survtmle.

ntrt
The number of trt values of interest.

uniqtrt
The unique values of trtOfInterest passed to mean_tmle.

t0
The timepoint at which survtmle was called to evaluate.

bounds
Minimum and maximum values to be placed on the ftype.

...
Other arguments. Not currently used.

Value

A list of data.frame objects as described above.