Mock RTSS/AS01 data set

A dataset containing data that is similar in structure to the RTSS/AS01 malaria vaccine trial. Privacy agreements prevent the sharing of the real data, so please note THAT THIS IS NOT THE REAL RTSS,S DATA. The data set is a list of 10 simulated multiple outputation draws. The covariate data, ftime, and vaccine stay the same across the data sets; however, the ftype variable changes, simulating outputed data sets of multiply infected trial participants.

rtss

Format

A list with 10 entries, each a data.frame with 6,890 rows and 18 columns.

ftime

number of months until first recorded malaria disease

ftype

the genotype of sampled malaria parasite (0 = censored, 1 = CSP matched, 2 = CSP mismatched)

vaccine

vaccine assignment (0 = control vaccine, 1 = vaccine)

ageWeeks

participant's age in weeks at trial enrollment

weightForAgeZscore

WHO weight-for-age Z-score

sex

participant's sex (0 = male, 1 = female)

site1-5

Indicator of study site

heightForAgeZscore

WHO height-for-age Z-score

weightForHeightZscore

WHO weight-for-height Z-score

armCircumZscore

WHO arm circumference Z-score

hemog

hemoglobin

distInpatient

distance from nearest inpatient clinic

distOutpatient

distance from nearest outpatient clinic

startMonthCat

study site-specific indicator of rainy (=1) vs. dry (=0) season
...